+++ Scripts/ParseEmailFiles/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-
+Added new argument *ignore_content_type* - Use it when the file has no Content-Type header. The script will treat the file as text
+    email with no attachments.
 
 ## [19.12.1] - 2019-12-25
 Added handling for cases where an attachment has neither the *DisplayName* nor the *AttachFilename* properties.