+++ Scripts/ParseEmailFiles/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+
+## [20.2.0] - 2020-02-04
 Added handling for EML files with no Content-Type header. The script will treat the file as email text with no attachments.
 
 ## [19.12.1] - 2019-12-25